% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{bic.func}
\alias{bic.func}
\title{Internal function for calculating the sBIC}
\usage{
bic.func(z, dat, uo, time, inflection.threshold, k)
}
\description{
Internal function for calculating the sBIC
}
