% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get.likelihood}
\alias{get.likelihood}
\title{Internal function for calculating the data likelihood given an assumed time-to-event density. Used by bic.func().}
\usage{
get.likelihood(dat, uo, time)
}
\description{
Internal function for calculating the data likelihood given an assumed time-to-event density. Used by bic.func().
}
