% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get.smooth.o}
\alias{get.smooth.o}
\title{Internal function for applying the Nadaraya–Watson kernel regression smoother to a time-to-event distribution.}
\usage{
get.smooth.o(z, turnbull.o, time)
}
\description{
Internal function for applying the Nadaraya–Watson kernel regression smoother to a time-to-event distribution.
}
