% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get.surv}
\alias{get.surv}
\title{Internal function for converting a time-to-event distribution to a survival distribution.}
\usage{
get.surv(uo)
}
\description{
Internal function for converting a time-to-event distribution to a survival distribution.
}
