% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{get.tp}
\alias{get.tp}
\title{Internal function for counting the number of turning points in a time-to-event curve. Used by bic.func().}
\usage{
get.tp(uo.smooth, inflection.threshold)
}
\description{
Internal function for counting the number of turning points in a time-to-event curve. Used by bic.func().
}
